Allow for multiple properties of same type
This allows you to, for instance, include several phone numbers (or
states, if you want, although that makes much less sense).

In order to do this, the data format has been radically changed: Instead
of depending on array keys for type, now it depends on, essentially, a
sort of de-facto tuple. Each row of data uses the format:

```
// old
['state' => 'OR']

// new
['state', 'OR']
```

This makes moving through arrays a bit easier in many ways (i.e.
array_map is useful now), but mostly it allows us to do this:

```
Scheme::build([
	['phone' => '(123) 456-7890'],
	['phone' => '(098) 765-4321'],
]);
```