Add zero-width space after inline elements
This should resolve issue #1.

Instead of adding a visual space, I've opted to add zero-width spaces.
These will allows the browser to break when appropriate, but will not
add any visual space--this is still left to CSS implementation.

In an idea world, we'd use the `<wbr>` element instead, but
unfortunately it isn't supported by (apparently) any version of IE,
which makes that a non-viable solution at this time.